Gamepad tweaks.

+++ firmware/sources/hardware/gamepads/Gamepad081FE401.h
@@ -1,11 +1,21 @@
+// ***************************************************************************************
+// ***************************************************************************************
+//
+//      Name :      Gamepad081FE401.h
+//      Authors :   Sascha Schneider
+//      Date :      25th March 2024
+//      Reviewed :  No
+//      Purpose :   Gamepad header (Olimex SNES style gamepad)
+//
+// ***************************************************************************************
+// ***************************************************************************************
+
 #pragma once
 #include "Gamepad.h"
 
-// Olimex SNES style gamepad
-
 class Gamepad081FE401 : public Gamepad {
 public:
-	uint8_t getState() override;
+	uint32_t getState() override;
 	void update(uint8_t dev_addr, uint8_t instance, uint8_t const* report, uint16_t len) override;
 
 private:
